+++ app/renderer/actions.js
@@ -50,20 +50,20 @@ const {
     currentMode,
     currentPage,
     currentTab,
-    getStored,
     getUrl,
     listRealPages,
     listTabs,
     sendToPageOrSubFrame,
-    setStored,
     tabForPage,
     updateGuiVisibility
 } = require("./common")
 
-let lastSearchFull = false
-let currentSearch = ""
-let storedSearch = ""
+let lastSearchLocal = true
+let lastSearchInc = true
+let lastIncSearch = ""
+let lastProperSearch = ""
 let searchDirection = "forward"
+let globalSearch = ""
 let potentialNewSearchDirection = "forward"
 
 /**
@@ -80,9 +80,17 @@ const emptySearch = args => {
     }
     if (["both", "global"].includes(scope)) {
         pages = listRealPages()
-        setStored("globalsearch", "")
+        globalSearch = ""
     }
-    pages.forEach(page => page?.stopFindInPage("clearSelection"))
+    pages.forEach(page => {
+        tabForPage(page)?.removeAttribute("search-results")
+        page?.stopFindInPage("clearSelection")
+    })
+    const searchResultsEl = document.getElementById("search-results")
+    if (searchResultsEl) {
+        searchResultsEl.style.display = "none"
+        searchResultsEl.textContent = ""
+    }
 }
 
 /**
@@ -99,26 +107,37 @@ const matchCase = search => {
 /** Highlight the next search match based on the scope. */
 const nextSearchMatch = () => {
     const scope = getSetting("searchscope")
-    let search = getStored("globalsearch")
+    let search = globalSearch
     let pages = []
     if (scope === "local") {
-        search = currentTab()?.getAttribute("localsearch")
-            || getStored("globalsearch")
+        search = currentTab()?.getAttribute("localsearch") || search
         pages = [currentPage()]
     } else {
         pages = listRealPages()
     }
-    if (search) {
-        pages.forEach(page => {
-            if (tabForPage(page)?.classList.contains("visible-tab")) {
-                page?.findInPage(search, {
-                    "findNext": true,
-                    "forward": searchDirection === "forward",
-                    "matchCase": matchCase(search)
-                })
+    if (!search) {
+        return
+    }
+    /** @type {"searchreachtraverse"|"searchreach"} */
+    let reachSetting = "searchreach"
+    if (getSetting("searchreachtraverse") !== "same") {
+        reachSetting = "searchreachtraverse"
+    }
+    const reach = getSetting(reachSetting)
+    if (reach === "all") {
+        pages = listRealPages()
+    } else {
+        pages = pages.filter(p => {
+            if (reach === "current") {
+                return p === currentPage()
             }
+            return p?.classList.contains("visible-page")
         })
     }
+    pages.forEach(p => p?.findInPage(search, {
+        "forward": searchDirection === "forward",
+        "matchCase": matchCase(search)
+    }))
 }
 
 /**
@@ -128,12 +147,11 @@ const nextSearchMatch = () => {
 const toSearchMode = args => {
     const {setMode} = require("./modes")
     setMode("search")
-    let search = getStored("globalsearch")
+    let search = globalSearch
     if (getSetting("searchscope") !== "global") {
-        search = currentTab()?.getAttribute("localsearch")
-            || getStored("globalsearch")
+        search = currentTab()?.getAttribute("localsearch") || search
     }
-    storedSearch = search
+    lastProperSearch = search
     if (args.hadModifier) {
         potentialNewSearchDirection = "backward"
     } else {
@@ -151,36 +169,37 @@ const toSearchMode = args => {
 /** Highlight the previous search match based on the scope. */
 const previousSearchMatch = () => {
     const scope = getSetting("searchscope")
-    let search = getStored("globalsearch")
+    let search = globalSearch
     let pages = []
     if (scope === "local") {
-        search = currentTab()?.getAttribute("localsearch")
-            || getStored("globalsearch")
+        search = currentTab()?.getAttribute("localsearch") || search
         pages = [currentPage()]
     } else {
         pages = listRealPages()
     }
-    if (search) {
-        pages.forEach(page => {
-            if (tabForPage(page)?.classList.contains("visible-tab")) {
-                page?.findInPage(search, {
-                    "findNext": true,
-                    "forward": searchDirection === "backward",
-                    "matchCase": matchCase(search)
-                })
+    if (!search) {
+        return
+    }
+    /** @type {"searchreachtraverse"|"searchreach"} */
+    let reachSetting = "searchreach"
+    if (getSetting("searchreachtraverse") !== "same") {
+        reachSetting = "searchreachtraverse"
+    }
+    const reach = getSetting(reachSetting)
+    if (reach === "all") {
+        pages = listRealPages()
+    } else {
+        pages = pages.filter(p => {
+            if (reach === "current") {
+                return p === currentPage()
             }
+            return p?.classList.contains("visible-page")
         })
     }
-}
-
-/**
- * Reset the incremental search match.
- * @param {ActionParam} args
- */
-const resetIncrementalSearch = args => {
-    if (getSetting("searchscope") === "inclocal" && !lastSearchFull) {
-        emptySearch({"scope": "local", "src": args?.src ?? "other"})
-    }
+    pages.forEach(p => p?.findInPage(search, {
+        "forward": searchDirection === "backward",
+        "matchCase": matchCase(search)
+    }))
 }
 
 /**
@@ -189,9 +208,11 @@ const resetIncrementalSearch = args => {
  */
 const incrementalSearch = args => {
     let scope = getSetting("searchscope")
+    const isIncSearch = args?.value === undefined
+    lastSearchLocal = scope === "local"
     if (scope === "inclocal") {
-        lastSearchFull = Boolean(args?.value)
-        if (args?.value === undefined) {
+        if (isIncSearch) {
+            lastSearchLocal = true
             scope = "local"
         } else {
             scope = "global"
@@ -199,29 +220,69 @@ const incrementalSearch = args => {
     }
     const url = getUrl()
     const search = args?.value || url?.value || ""
+    if (lastProperSearch === search) {
+        if (!lastSearchInc && !isIncSearch) {
+            nextSearchMatch()
+            return
+        }
+        if (lastIncSearch === search && isIncSearch) {
+            return
+        }
+    }
+    lastIncSearch = search
+    if (!isIncSearch) {
+        lastProperSearch = search
+    }
+    lastSearchInc = isIncSearch
     let pages = [currentPage()]
     if (scope === "global") {
         pages = listRealPages()
-        setStored("globalsearch", search)
+        globalSearch = search
+        listTabs().forEach(t => t.removeAttribute("localsearch"))
     } else {
         currentTab()?.setAttribute("localsearch", search)
     }
-    if (search === currentSearch) {
-        if (args && search === storedSearch) {
-            nextSearchMatch()
-        }
-        return
+    /** @type {"searchreachinc"|"searchreach"} */
+    let reachSetting = "searchreach"
+    if (isIncSearch && getSetting("searchreachinc") !== "same") {
+        reachSetting = "searchreachinc"
     }
-    currentSearch = search
-    if (search) {
-        pages.forEach(page => {
-            page?.stopFindInPage("clearSelection")
-            if (tabForPage(page)?.classList.contains("visible-tab")) {
-                page?.findInPage(search, {"matchCase": matchCase(search)})
+    const reach = getSetting(reachSetting)
+    if (reach === "all") {
+        pages = listRealPages()
+    } else {
+        pages = pages.filter(p => {
+            if (reach === "current") {
+                return p === currentPage()
             }
+            return p?.classList.contains("visible-page")
         })
+    }
+    pages.forEach(p => p?.stopFindInPage("clearSelection"))
+    if (search) {
+        pages.forEach(p => p?.findInPage(search, {
+            "findNext": true, "matchCase": matchCase(search)
+        }))
     } else {
         emptySearch({scope, "src": args?.src ?? "other"})
+    }
+}
+
+/**
+ * Reset the incremental search match.
+ * @param {ActionParam} args
+ */
+const resetIncrementalSearch = args => {
+    if (getSetting("searchscope") === "inclocal" && lastSearchLocal) {
+        emptySearch({"scope": "local", "src": args?.src ?? "other"})
+    } else if (getSetting("incsearch") && lastProperSearch !== lastIncSearch) {
+        if (lastProperSearch) {
+            incrementalSearch({
+                "src": args.src ?? "other", "value": lastProperSearch
+            })
+        } else {
+            emptySearch({"scope": "both", "src": args?.src ?? "other"})
+        }
     }
 }
 
@@ -959,12 +1020,8 @@ const toNormalMode = () => {
 
 /** Go to the previous mode used before follow, or back to normal. */
 const stopFollowMode = () => {
-    const {setMode} = require("./modes")
-    if (currentMode() === "follow") {
-        setMode(getStored("modebeforefollow") || "normal")
-    } else {
-        setMode("normal")
-    }
+    const {stopFollow} = require("./follow")
+    stopFollow()
 }
 
 /**